requirements.txt
accelerate==0.29.1
aiohttp==3.9.3
aiosignal==1.3.1
arrow==1.3.0
async-timeout==4.0.3
attrs==23.2.0
boto3==1.34.79
botocore==1.34.79
bravado==11.0.3
bravado-core==6.1.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
datasets==2.18.0
dill==0.3.8
filelock==3.13.3
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.2.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.22.2
idna==3.6
isoduration==20.11.0
isort==5.13.2
Jinja2==3.1.3
jmespath==1.0.1
jsonpointer==2.4
jsonref==1.1.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
monotonic==1.6
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
neptune==1.10.1
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
packaging==24.0
pandas==2.2.1
pillow==10.3.0
psutil==5.9.8
pyarrow==15.0.2
pyarrow-hotfix==0.6
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
s3transfer==0.10.1
safetensors==0.4.2
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
swagger-spec-validator==3.0.3
sympy==1.12
tokenizers==0.15.2
torch==2.2.2
tqdm==4.66.2
-e git+https://github.com/szysad/transformers@f633bdc5224b0102289f3f97b5269a31a828f4dc#egg=transformers
triton==2.2.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.1
webcolors==1.13
websocket-client==1.7.0
xxhash==3.4.1
yarl==1.9.4
